refactor: RedisUtil 클래스를 레포지토리로 이동 및 클래스명 변경
